+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- Show InfluxDB write URL on startup instead of InfluxDB host
 - Remove deprecated `topic-1` style tags. (deprecated since 0.4.0)
 
 ## [1.0.0] - 2024-02-16